requirements.txt
discord==2.3.2
discord.py==2.4.0
python-dotenv==1.0.1
selenium==4.27.1